refactor: remove global function
